pyproject.toml
[project]
name = "reliquary-inference"
version = "0.1.0"
description = "Reliquary Ledger — proof-carrying inference runtime for Bittensor (testnet netuid 462)"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Reliquary contributors" }]
keywords = [
  "bittensor",
  "subnet",
  "reliquary",
  "inference",
  "proof-of-inference",
  "grail",
  "dapo",
  "grpo",
  "math-reasoning",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "aiobotocore>=2.13.1",
  "bittensor>=9.12.1",
  "bittensor-wallet>=4.0.0",
  "datasets>=2.20.0",
  "numpy>=1.26.0",
  "pyarrow>=16.1.0",
  "rich>=13.7.1",
  "safetensors>=0.4.3",
  "torch>=2.4.0",
  "transformers>=4.52.0",
  "typer>=0.12.3",
]

[project.urls]
Homepage = "https://github.com/0xgrizz/reliquary-inference"
Repository = "https://github.com/0xgrizz/reliquary-inference"
Issues = "https://github.com/0xgrizz/reliquary-inference/issues"

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "hypothesis>=6.100.0",
]
gpu = [
  "vllm>=0.8.5",
]
tracing = [
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
]
protocol = [
  "reliquary-protocol>=0.1.0",
]
r2 = [
  "boto3>=1.30.0",
]

[project.scripts]
reliquary-inference = "reliquary_inference.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
# Test files + the closed-loop-e2e fixtures do sys.path manipulation
# (to inject sibling-repo imports) BEFORE importing — ruff E402 can't
# tell the difference.
"tests/**/*.py" = ["E402", "E741"]

[tool.pytest.ini_options]
testpaths = ["tests"]
